Refactor: Moved the EditRecipeScreenViewModel to a new package, and modified the EditRecipeScreen
The `EditRecipeScreenViewModel` was removed from its previous package and placed under a new directory.
The `EditRecipeScreen` was modified to interact with the new `EditRecipeScreenViewModel`.
Added new category dialog.
Modified input field to accept only numbers.
Added new text to explain the ingredients and cooking steps.
Added logic to initialize with existing recipe or not.
Added button logic to show add or edit depending on the existing recipe.